[scripts/wastickers] Rewrote in Python
This rewrite was nessecary, since, apparently, the format of the
sticker packs that WhatsApp reads is a bit... weird. All files start with a
leading slash, which the default `zip` program strips before packing.
The `zipfile` module in Python does not have this check, so it was a nessecary sacrifice.